refactor(storage): split connect.go into explicit bootstrap phases
Split internal/storage/connect.go, which mixed four concerns behind one
Open entry point, into one file per concern:

- open.go: openDB builds the DSN and documents every pragma one for one
- migrate.go: NewMigrationProvider and runMigrations (the issue #691
  registration guard lives here)
- credentials.go: keyring namespace scoping and credential-location
  bookkeeping
- heal.go: the declared healSteps slice and the heal runner
- heal_alarm_uids.go, heal_xprops.go, heal_alarm_repeat.go: one heal
  pass per file

Open is now composition only. The heal order is visible in the declared
slice, and a table-driven test pins the step names and asserts that
every step is idempotent. No behavior change: same pragmas, same
migration set, same heal order, same error messages.

Closes #735
